Broadcast consider data
Whenever the plugin considers a room, broadcast the results to all other plugins. This will allow other scripts to do things like count mobs, alert the user depending on whether certain mob names are present, and much more.

Use JSON instead of the serialize module for easier and more widespread use. The plugin's config data is still (de-)serialized using serialize to avoid forcing users to rebuild their individual config settings.

Update `Consider.start()` to return true/false to indicate whether a coroutine was started. Plugin authors will use this to kick off a consider request and tell whether it is successful.

Update `Consider.startCR()` to broadcast empty data in the event of a timeout so other plugins don't get hung up waiting.

Add explanation and example of broadcast-related info to the 'Consider' helpfile. This may not be the ideal spot for it, but lazy...