Disable auto-con when zone changes
Rezit suggested disabling auto-con when entering certain areas to cut down on spam. Because determining which areas to include would be fairly subjective, and because the list would change as new areas are added anyway, decide to disable auto-con whenever the user moves to a different zone, forcing them to re-enable it if they want. This will, hopefully, cut down on spam caused by inattentive/inconsiderate users.